Move common code from berlinger module out into lib module Modify lib functions to handle both sensor types (LogTag files are .csv and start with "LogTag")
