+++ src/berlinger.rs
@@ -1,13 +1,9 @@
 use chrono::{Duration, NaiveDate, NaiveDateTime, NaiveTime};
 use serde_json::{json, Value};
-use std::fs;
 use std::fs::File;
 use std::io;
 use std::io::BufRead;
 use std::path::Path;
-
-#[cfg(any(target_os = "windows", target_os = "linux"))]
-use rs_drivelist::drive_list;
 
 use crate::common::{
     BreachType, Sensor, SensorType, TemperatureBreach, TemperatureBreachConfig, TemperatureLog,
@@ -777,164 +773,6 @@ pub fn read_sensor_from_file(file_path: &str) -> Option<Sensor> {
         Some(sensor)
     } else {
         log::error!("File not found: {}", file_path);
-        None
-    }
-}
-
-#[cfg(target_os = "macos")]
-fn sensor_volume_paths() -> Vec<String> {
-    let mut volume_list: Vec<String> = Vec::new();
-
-    if let Ok(entries) = fs::read_dir("/Volumes") {
-        // loop over folders in Volumes
-        for entry in entries {
-            if let Ok(entry) = entry {
-                if entry.path().is_dir() {
-                    if let Some(txt_file_path) = entry.path().to_str() {
-                        volume_list.push(txt_file_path.to_string())
-                    }
-                }
-            }
-        }
-    }
-    volume_list
-}
-
-#[cfg(target_os = "android")]
-fn sensor_volume_paths() -> Vec<String> {
-    let mut volume_list: Vec<String> = Vec::new();
-
-    if let Ok(entries) = fs::read_dir("/mnt/media_rw") {
-        // loop over mounted media folders
-        for entry in entries {
-            if let Ok(entry) = entry {
-                if entry.path().is_dir() {
-                    if let Some(txt_file_path) = entry.path().to_str() {
-                        volume_list.push(txt_file_path.to_string())
-                    }
-                }
-            }
-        }
-    }
-    volume_list
-}
-
-#[cfg(any(target_os = "windows", target_os = "linux"))]
-fn sensor_volume_paths() -> Vec<String> {
-    let mut volume_list: Vec<String> = Vec::new();
-
-    match drive_list() {
-        Err(err) => log::error!("No drives found: {}", err),
-        Ok(drives) => {
-            for drive_index in 0..drives.len() {
-                // loop over all detected drives
-
-                let mount_points = &drives[drive_index].mountpoints;
-                for partition_index in 0..mount_points.len() {
-                    // loop over partitions
-                    let mount_point = &mount_points[partition_index];
-
-                    if mount_point.totalBytes < Some(8 * 1024 * 1024 * 1024) {
-                        // possible USB drive if < 8 GB
-                        volume_list.push(mount_point.path.clone());
-                    }
-                }
-            }
-        }
-    }
-
-    volume_list
-}
-
-fn sensor_file_list() -> Vec<String> {
-    let mut file_list: Vec<String> = Vec::new();
-
-    for volume_root in sensor_volume_paths() {
-        // loop over volumes
-
-        if let Ok(entries) = fs::read_dir(&volume_root) {
-            // loop over files in the volume root
-            for entry in entries {
-                if let Ok(entry) = entry {
-                    if let Some(extension) = entry.path().extension() {
-                        if extension == "txt" {
-                            // might be a sensor txt file
-                            if let Some(txt_file_path) = entry.path().to_str() {
-                                let pdf_file_path = txt_file_path.replace(".txt", ".pdf");
-
-                                if Path::new(&pdf_file_path).exists() {
-                                    // but only if it has a matching PDF
-                                    file_list.push(txt_file_path.to_string())
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    file_list
-}
-
-fn sensor_serial_from_file_path(txt_file_path: &str) -> Option<String> {
-    let mut valid_serial = false;
-    let mut serial = "";
-    let file_path = Path::new(&txt_file_path);
-
-    if file_path.exists() {
-        if let Some(os_file_name) = file_path.file_name() {
-            if let Some(file_name) = os_file_name.to_str() {
-                let elements: Vec<&str> = file_name.split("_").collect();
-                serial = elements[0];
-                valid_serial = true;
-            }
-        }
-    }
-
-    if valid_serial {
-        Some(serial.to_string())
-    } else {
-        None
-    }
-}
-
-/// Returns all the serials found from currently mounted USB drives up to 8GB capacity
-/// (-> any USB drive containing sensor files if you don't have a physical sensor).
-/// For Berlinger sensors, it expects to find a serial_xxxxx.txt file in the root folder
-/// together with a matching PDF file (USB drives can have multiple pairs of files).
-pub fn read_sensor_serials() -> Option<Vec<String>> {
-    let mut serial_list: Vec<String> = Vec::new();
-
-    for txt_file_path in sensor_file_list() {
-        if let Some(serial) = sensor_serial_from_file_path(&txt_file_path) {
-            serial_list.push(serial)
-        }
-    }
-
-    if serial_list.len() > 0 {
-        Some(serial_list)
-    } else {
-        None
-    }
-}
-
-/// Returns all sensors found from currently mounted USB drives up to 8GB capacity
-/// (-> any USB drive containing sensor files if you don't have a physical sensor).
-/// For Berlinger sensors, it expects to find a serial_xxxxx.txt file in the root folder
-/// together with a matching PDF file (USB drives can have multiple pairs of files).
-pub fn read_sensors_from_usb() -> Option<Vec<Sensor>> {
-    let mut sensors: Vec<Sensor> = Vec::new();
-
-    for txt_file_path in sensor_file_list() {
-        if let Some(sensor) = read_sensor_from_file(&txt_file_path) {
-            sensors.push(sensor.clone())
-        }
-    }
-
-    if sensors.len() > 0 {
-        Some(sensors)
-    } else {
         None
     }
 }
